requirements.txt
chromadb==0.5.5
langchain-chroma==0.1.2
langchain==0.2.11
langchain-community==0.2.10
langchain-text-splitters==0.2.2
langchain-groq==0.1.6
transformers==4.43.2
sentence-transformers==3.0.1
unstructured==0.15.0
unstructured[pdf]==0.15.0
pyarrow
streamlit-option-menu
google-generativeai
ipython
pysqlite3